backend/pyproject.toml
[tool.poetry]
name = "ai-presentation-intelligence-platform"
version = "0.1.0"
description = "AI Presentation Intelligence Platform Backend"
authors = ["Platform Team"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
redis = {extras = ["hiredis"], version = "^5.0.4"}
celery = {extras = ["redis"], version = "^5.4.0"}
langchain = "^0.2.0"
langchain-anthropic = "^0.1.15"
langchain-openai = "^0.1.8"
langchain-groq = "^0.1.5"
langsmith = "^0.1.77"
sentence-transformers = "^3.0.0"
python-pptx = "^0.6.23"
boto3 = "^1.34.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
httpx = "^0.27.0"
tenacity = "^8.3.0"
structlog = "^24.2.0"
jsonschema = "^4.22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
hypothesis = "^6.103.0"
httpx = "^0.27.0"
factory-boy = "^3.3.0"
ruff = "^0.4.7"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
